* applied several patches from Bugtracker (thanks jens and others...) - handle special charakters in tokenizer (fixes freeze) - fix some compiler warnings on windows (auto-import) - fixed issues with wxSmith on 64bit systems - choose the right target for debugging - correct placement of file modified dialog - fixed wrong behaviour of taskbar icon on batch builds - save missing layout infos
git-svn-id: https://svn.code.sf.net/p/codeblocks/code/trunk@5117 2a5c6006-c6dd-42ca-98ab-0921f2732cef

+++ src/plugins/contrib/wxSmith/wxwidgets/defitems/wxschoice.h
@@ -39,7 +39,7 @@ class wxsChoice: public wxsWidget
         virtual void OnEnumWidgetProperties(long Flags);
 
         wxArrayString ArrayChoices;
-        int DefaultSelection;
+        long DefaultSelection;
 };
 
 #endif
